shell-player: calculate duration properly when not supplied by the player
Another gint64 cast failure.  This was breaking playback of ogg vorbis
over HTTP, amongst other things.

+++ shell/rb-shell-player.c
@@ -3494,7 +3494,7 @@ tick_cb (RBPlayer *mmplayer,
 	 * value from the entry, if any.
 	 */
 	if (duration < 1) {
-		duration = rhythmdb_entry_get_ulong (entry, RHYTHMDB_PROP_DURATION) * RB_PLAYER_SECOND;
+		duration = ((gint64)rhythmdb_entry_get_ulong (entry, RHYTHMDB_PROP_DURATION)) * RB_PLAYER_SECOND;
 		duration_from_player = FALSE;
 	}
